fix(scenario-convert): improve export of MOSAIC scenarios (#488)
* feat(utils): provide extension for FileIncrement
* fix(cell): make capacity nullable, as it is not wanted/required for server configurations
* fix(utils): only write 5 decimal places at most when exporting unit fields
* fix(traffic-signs): configure visibility correctly (clamp was use incorrectly)
* fix(mapping): handle new null case for parameter deviations

+++ fed/mosaic-cell/src/test/java/org/eclipse/mosaic/fed/cell/config/CNetworkImportTest.java
@@ -124,9 +124,9 @@ public class CNetworkImportTest {
 
         assertEquals(0.1, server.downlink.unicast.transmission.lossProbability, EPSILON);
 
-        assertEquals(Long.MAX_VALUE, server.uplink.capacity);
+        assertEquals(Long.MAX_VALUE, server.uplink.capacity.longValue());
         assertEquals(Long.MAX_VALUE, server.uplink.maxCapacity);
-        assertEquals(Long.MAX_VALUE, server.downlink.capacity);
+        assertEquals(Long.MAX_VALUE, server.downlink.capacity.longValue());
         assertEquals(Long.MAX_VALUE, server.downlink.maxCapacity);
     }
 }
